RJ-P5 (seal core): PG counts.json + guard-clean partition model + live two-engine seal
The engine + conform half of the seal, proven live on the RH-1 returns_ingest
program over 🐳 PG16 + Polars 1.42.1 (eight/eight green, rejects/bad streams
byte-equal as multisets, in==processed+rejects balancing identically, canary red).

- Partition model fix: `processed` is now counted at the guard's CLEAN output
  (branch-true child), not the sibling OUT ports. The old model was RH-1-
  coincidental (guard → row-preserving branch); a guard followed by a row-
  dropping/-collapsing op (the hero's join+aggregate) would diverge. Shared
  `RejectSites` resolver (emit/core) anchors both emitters to the same nodes so
  the two engines' counts.json triples agree by construction.
- PG counts.json emit: `SqlIslandEmitter.countQueries` builds three independent
  `count(*)` cones per site (in / clean-output / reject terminal) via
  `SqlGraphEmitter.planForNode`; `PgAdbcIslandEmitter` writes counts.json on the
  same adbc session. Rejects computed over the ACTUAL terminal so a broken
  producer imbalances the triple (canary caught on PG as on Polars).
- Clean-value canonicalization: the guarded clean cast strips ASCII-ws then casts
  on Polars, and canonicalizes int→bigint on PG (the int64 guard domain — a value
  in (int32max, int64max] passes the guard, so `AS integer` would overflow; also
  matches Polars pl.Int64). Cross-engine clean values + schema now conform, not
  just counts. Gated on a guarded plan ⇒ fail-fast twin stays byte-identical (R-P3).
- Conform: rejects/bad staging streams are first-class Q9 compare targets (5.1.3),
  paired by port name from the manifest reject sites (multiset, schema incl.
  `_ttrp_reject_*`). New gated `RejectsConformLiveTest` (5.2.4/5.2.6) + reject CSV.

Grammar has no `rejects` port kind (portDecl: IN|OUT|ERR) — the hero's `err
rejects` matches RH-1, so task 5.2.1(c)'s premise is void: no header change at
re-author time. Offline emit/cli/conform/graph suites green; my-module ktlint clean.

+++ packages/kotlin/ttrp-emit/src/test/golden/sql/postgres/rejects_pg_bad.sql
@@ -8,7 +8,7 @@ WITH "checked_guard" AS (
   WHERE "_ttrp_v1"
 ),
 "checked_1" AS (
-  SELECT "customer", "region", "amount", (CAST("customer" AS integer)) AS "returned_qty"
+  SELECT "customer", "region", "amount", (CAST("customer" AS bigint)) AS "returned_qty"
   FROM "checked_branch_t"
 )
 SELECT *
